getAll ...

+++ remote.sh
@@ -1,6 +1,14 @@
 #!/bin/sh
 # Created with /Users/ivo/Dropbox/Shell-Scripts/cmd/crea at 2022-05-01 08:23:57
 dir=$(CDPATH= cd -- "$(dirname -- "$0")" && pwd)
+busy="$(readlink /bin/ls)"
+if [ "$busy" = "busybox" ]; then 
+	c=""
+	n=-n
+else
+	c="\c"
+	n=""
+fi
 bridge=0
 if [[  "$1" =~ ^[0-9]{1} ]]; then
 	bridge="$1"
@@ -17,7 +25,6 @@ if [ ! -f "$config" ]; then
 	echo "The bridge '$bridge' is not available"!
 	exit 1
 fi
-
 ClientId=fv8jJ0lr3BrqyAuu9Vmxpvte7qNyqDn0
 ClientSecret=HuvOrVmAEnQfJ9nk
 appid=plug-switcher
@@ -25,7 +32,7 @@ Host="https://api.meethue.com"
 script="remote.sh"
 
 encode() {
-	echo "$1\c" | curl -Gso /dev/null -w %{url_effective} --data-urlencode @- "" | cut -c 3-
+	echo $n "$1$c" | curl -Gso /dev/null -w %{url_effective} --data-urlencode @- "" | cut -c 3-
 }
 testToken () {
 	if [ -f "$env" ]; then
@@ -52,7 +59,7 @@ testToken () {
 	fi
 	if [ -z "$access_token" ]; then 
 		echo "access_token not found!"
-		err=1
+		err=2
 	fi
 	if [ -z "$refresh_token" ]; then 
 		echo "refresh_token not found!"
@@ -68,6 +75,8 @@ testToken () {
 	fi
 }
 
+### gettoken
+
 what=$(echo "$1"  | tr '[:upper:]' '[:lower:]')
 if [ "$what" = "gettoken" ]; then
 	devicename="$(hostname | cut -f 1 -d '.' )"
@@ -76,32 +85,47 @@ if [ "$what" = "gettoken" ]; then
 	fi
 	deviceid=$(echo $devicename | tr " " "_" | tr '[:upper:]' '[:lower:]')
 	response_type=code
-	state=$(date -v-1H -u +fiveyears%Y%m%d%H | tr -d '\n' | md5)
+	if [ "$busy" = "busybox" ]; then 
+		state=$(($(date -u +%s) - 3660))  
+		state=$(date -d $state -u -D %s +fiveyears%Y%m%d%H   | tr -d '\n' | openssl md5 )
+		state=$(echo ${state/* })
+	else
+		state=$(date -v-1H -u +fiveyears%Y%m%d%H | tr -d '\n' | md5)
+	fi
 	deviceid=$(encode $deviceid)
 	devicename=$(encode $devicename)
 	code_verifier=$(openssl rand -base64 66 | tr -d '\n' | sed 's/+/_/g' | sed 's/\//_/g' | sed 's/=//g')
-	code_challenge=$(echo "$code_verifier\c" | openssl sha256 -binary | base64 | tr '/+' '_-' | tr -d '=')
+	code_challenge=$(echo $n "$code_verifier$c" | openssl sha256 -binary | base64 | tr '/+' '_-' | tr -d '=')
 	code_challenge_method=S256
 	filename="hue_${state}.txt"
 	uri="/v2/oauth2/authorize"
 	rm -f "$HOME/Downloads/"hue_*.txt 2>/dev/null
-	open "$Host$uri?client_id=$ClientId&response_type=$response_type&state=$state&appid=$appid&deviceid=$deviceid&devicename=$devicename&code_challenge_method=$code_challenge_method&code_challenge=$code_challenge"
-	echo "Wait 10 seconds for '$filename' \c"
-	i=0
-	while [[ ! -e "$HOME/Downloads/$filename" ]] ; do
-		((i++))
-		if [[ $i -gt 10 ]]; then
-			echo
-			echo "'$filename' not found!"
-			echo "Please start over!"
-			exit 1
-		fi
-		echo ".\c"
-	    sleep 1
-	done
-	echo -e "\033[2K"
-	code=$(cat "$HOME/Downloads/$filename")
-	rm -f "$HOME/Downloads/"hue_*.txt 2>/dev/null
+	if [ "$busy" = "busybox" ]; then 
+		echo "Copy url in Browser:"
+	    echo "$Host$uri?client_id=$ClientId&response_type=$response_type&state=$state&appid=$appid&deviceid=$deviceid&devicename=$devicename&code_challenge_method=$code_challenge_method&code_challenge=$code_challenge"
+	    echo
+	    echo "Open downloaded file \"hue_${state}.txt\" and copy code!"
+	    echo
+	    read -p "Put in code: " code
+	else
+		open "$Host$uri?client_id=$ClientId&response_type=$response_type&state=$state&appid=$appid&deviceid=$deviceid&devicename=$devicename&code_challenge_method=$code_challenge_method&code_challenge=$code_challenge"
+		echo $n "Wait 20 seconds for '$filename' $c"
+		i=0
+		while [[ ! -e "$HOME/Downloads/$filename" ]] ; do
+			i=$((i+1))
+			if [[ $i -gt 20 ]]; then
+				echo
+				echo "'$filename' not found!"
+				echo "Please start over!"
+				exit 1
+			fi
+			echo $n ". $i $c"
+		    sleep 1
+		done
+		echo -e "\033[2K"
+		code=$(cat "$HOME/Downloads/$filename")
+		rm -f "$HOME/Downloads/"hue_*.txt 2>/dev/null
+	fi
 	if [ -z "$code" -o -z "$code_verifier" ]; then
 		echo "Please start authentication with $(basename $0) getCode."
 		exit 1
@@ -120,9 +144,18 @@ if [ "$what" = "gettoken" ]; then
 		echo "Please start over!"
 		exit 1
 	fi
-	HASH1=$(echo "$ClientId:$realm:$ClientSecret\c" | md5)
-	HASH2=$(echo "POST:$uri\c" | md5)
-	response=$(echo "$HASH1:$nonce:$HASH2\c" | md5)
+	if [ "$busy" = "busybox" ]; then 
+		HASH1=$(echo $n "$ClientId:$realm:$ClientSecret$c" | openssl md5 )
+		HASH1=$(echo ${HASH1/* })
+		HASH2=$(echo $n "POST:$uri$c" | openssl md5 )
+		HASH2=$(echo ${HASH2/* })
+		response=$(echo $n "$HASH1:$nonce:$HASH2$c" | openssl md5 )
+		response=$(echo ${response/* })
+	else
+		HASH1=$(echo $n "$ClientId:$realm:$ClientSecret$c" | md5)
+		HASH2=$(echo $n "POST:$uri$c" | md5)
+		response=$(echo $n "$HASH1:$nonce:$HASH2$c" | md5)
+	fi
 	ret=$(curl -s --request POST "$Host$uri" -d "grant_type=authorization_code&code=$code&code_verifier=$code_verifier" \
 		-H "Authorization: Digest username=\"$ClientId\", realm=\"$realm\", nonce=\"$nonce\", uri=\"$uri\", response=\"$response\"" \
 		-H 'Content-Type: application/x-www-form-urlencoded'  2>&1 )
@@ -137,8 +170,15 @@ if [ "$what" = "gettoken" ]; then
 	echo "access_token=$access_token" >> "$env"
 	echo "refresh_token=$refresh_token" >> "$env"
 	expires_at=$(($expires_in + $(date +%s)))
-	echo "expires_at=$expires_at\c" >> "$env"
-	echo "   # ($(date -j -f "%s" $expires_at "+%Y-%m-%d %H:%M:%S"))" >> "$env"
+	echo $n "expires_at=$expires_at$c" >> "$env"
+	if [ "$busy" = "busybox" ]; then 
+		echo "   # ($(date  -d $expires_at -D %s "+%Y-%m-%d %H:%M:%S"))" >> "$env"
+	else
+		echo "   # ($(date -j -f "%s" $expires_at "+%Y-%m-%d %H:%M:%S"))" >> "$env"
+	fi
+
+### user
+
 elif [ "$what" = "user" ]; then
 	user=$(cat $config | grep "set user" | grep -v default | tr -s ' ' | cut -d " " -f 3)
 	if [ -z $user ]; then
@@ -146,6 +186,9 @@ elif [ "$what" = "user" ]; then
 		user=$(cat $config | grep "set user" | grep -v default | tr -s ' ' | cut -d " " -f 3)
 	fi
 	echo $user
+
+### bluebutton
+
 elif [ "$what" = "bluebutton" ]; then
 	token=$("$0" token)
 	uri=/route/api/0/config 
@@ -154,6 +197,9 @@ elif [ "$what" = "bluebutton" ]; then
 		 --header 'Content-Type: application/json' \
 		 --data-raw '{ "linkbutton":true }' 
 	echo
+
+### newuser
+
 elif [ "$what" = "newuser" ]; then
 	token=$("$0" token)
 	uri=/route/api/0/config 
@@ -173,6 +219,9 @@ elif [ "$what" = "newuser" ]; then
 	echo $line1 >| $config 
 	echo $line2 >> $config
 	echo $line3 >> $config
+
+### deleteuser
+
 elif [ "$what" = "deleteuser" ]; then
 	token=$("$0" token)
 	user=$(cat $config | grep "set user" | grep -v default | tr -s ' ' | cut -d " " -f 3)
@@ -183,10 +232,16 @@ elif [ "$what" = "deleteuser" ]; then
 	line3=$(cat $config | grep -v "set user")
 	echo $line1 >| $config 
 	echo $line3 >> $config
+
+### newtoken
+
 elif [ "$what" = "newtoken" ]; then
 	testToken
 	"$0" refreshToken 1
 	exit 0
+
+### token
+
 elif [ "$what" = "token" ]; then
 	testToken
 	if [ $(($expires_at - $(date +%s) ))  -lt 3660 ]; then
@@ -194,26 +249,37 @@ elif [ "$what" = "token" ]; then
 		exit 0
 	fi
 	echo $access_token
+
+### checktoken
+
 elif [ "$what" = "checktoken" ]; then
 	testToken
 	echo "Access_token : $access_token"
-	echo "  expires at : $(date -j -f "%s" $expires_at "+%Y-%m-%d %H:%M:%S")"
+	echo -n "  expires at : "
+	if [ "$busy" = "busybox" ]; then 
+		echo "   # ($(date  -d $expires_at -D %s "+%Y-%m-%d %H:%M:%S"))" 
+	else
+		echo "   # ($(date -j -f "%s" $expires_at "+%Y-%m-%d %H:%M:%S"))" 
+	fi
 	echo "Refresh_token: $refresh_token"
 	if [ $(($expires_at - $(date +%s) ))  -lt 3660 ]; then
 		echo "Access_token is almost expired!"
-		echo "Please $0 refresh"
+		echo "Please $0 refreshToken"
 	fi
 	echo "code_verifier: $code_verifier"
 	echo "nonce        : $nonce"
 	echo "realm        : $realm"
 	echo "response     : $response"
+
+### refreshtoken
+
 elif [ "$what" = "refreshtoken" ]; then
 	# $1 = 1 -> Aufruf durch token
 	if [ -z "$2" ]; then
 		testToken
 	fi
 	uri="/v2/oauth2/token"
-	ret=$(curl -s --request POST "$Host$uri" -d "grant_type=refresh_token&refresh_token=$refresh_token&code_verifier=$code_verifier" \
+	ret=$(curl -s -S --request POST "$Host$uri" -d "grant_type=refresh_token&refresh_token=$refresh_token&code_verifier=$code_verifier" \
 		-H "Authorization: Digest username=\"$ClientId\", realm=\"$realm\", nonce=\"$nonce\", uri=\"$uri\", response=\"$response\"" \
 		-H 'Content-Type: application/x-www-form-urlencoded'  2>&1 )
 	ret=$(echo "$ret" | cut -f 4,7,10 -d "\"" | tr -d ': ,')
@@ -226,11 +292,18 @@ elif [ "$what" = "refreshtoken" ]; then
 	echo "access_token=$access_token" >> "$env"
 	echo "refresh_token=$refresh_token" >> "$env"
 	expires_at=$(($expires_in + $(date +%s)))
-	echo "expires_at=$expires_at\c" >> "$env"
-	echo "   # ($(date -j -f "%s" $expires_at "+%Y-%m-%d %H:%M:%S"))" >> "$env"
+	echo $n "expires_at=$expires_at$c" >> "$env"
+	if [ "$busy" = "busybox" ]; then 
+		echo "   # ($(date  -d $expires_at -D %s "+%Y-%m-%d %H:%M:%S"))"  >> "$env"
+	else
+		echo "   # ($(date -j -f "%s" $expires_at "+%Y-%m-%d %H:%M:%S"))"  >> "$env"
+	fi
 	if [ -n "$2" ]; then
 		echo $access_token
 	fi
+
+### help
+
 else
 	echo "Token commands"
 	echo "   $(basename $0) token | newToken | getToken | refreshToken | checkToken"
